+++ src/checks_list.cc
@@ -13,28 +13,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-#include "src/check_list.h"
+#include "src/checks_list.h"
 
 #include "absl/strings/string_view.h"
-#include "src/linter.h"
+#include "src/checks.h"
 #include "src/linter_options.h"
 
 namespace zetasql::linter {
 
 const std::vector<
     std::function<LinterResult(absl::string_view, const LinterOptions&)>>
-CheckList::GetList() {
+ChecksList::GetList() {
   return list_;
 }
 
-void CheckList::Add(
+void ChecksList::Add(
     std::function<LinterResult(absl::string_view, const LinterOptions&)>
         check) {
   list_.push_back(check);
 }
 
-CheckList GetParserDependantChecks() {
-  CheckList list;
+ChecksList GetParserDependantChecks() {
+  ChecksList list;
   list.Add(CheckSemicolon);
   list.Add(CheckAliasKeyword);
   list.Add(CheckNames);
@@ -42,8 +42,8 @@ CheckList GetParserDependantChecks() {
   return list;
 }
 
-CheckList GetAllChecks() {
-  CheckList list;
+ChecksList GetAllChecks() {
+  ChecksList list;
   list.Add(CheckLineLength);
   list.Add(CheckParserSucceeds);
   list.Add(CheckSemicolon);
