+++ src/checks_util.cc
@@ -24,7 +24,7 @@
 #include "absl/strings/str_cat.h"
 #include "absl/strings/str_split.h"
 #include "absl/strings/string_view.h"
-#include "src/lint_errors.h"
+#include "src/lint_error.h"
 #include "src/linter_options.h"
 #include "zetasql/parser/parse_tree.h"
 #include "zetasql/parser/parse_tree_visitor.h"
@@ -214,8 +214,7 @@ bool ConsistentUppercaseLowercase(const absl::string_view &sql,
   }
   // There shouldn't be any case any Keyword
   // contains both uppercase and lowercase characters
-  if (option.UpperKeyword())
-    return !lowercase;
+  if (option.UpperKeyword()) return !lowercase;
   return !uppercase;
 }
