+++ src/checks_util.h
@@ -25,7 +25,7 @@
 
 #include "absl/status/status.h"
 #include "absl/strings/string_view.h"
-#include "src/lint_errors.h"
+#include "src/lint_error.h"
 #include "src/linter_options.h"
 #include "zetasql/parser/parse_tree.h"
 #include "zetasql/parser/parse_tree_visitor.h"
