+++ src/lint_error.cc
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-#include "src/lint_errors.h"
+#include "src/lint_error.h"
 
 #include <algorithm>
 #include <cstdio>
